pyproject.toml
[tool.poetry]
name = "wovenlight-tfl"
version = "1.0.0"
description = "Schedules calls to TfL Line Disruption API and stores results."
authors = ["Syed Saad Amer <syedsaadamer@gmail.com>"]
readme = "README.md"
# ensure the 'app' package is included
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = ">=0.112"
uvicorn = {version = ">=0.30", extras = ["standard"]}
sqlalchemy = ">=2.0"
apscheduler = ">=3.10"
requests = ">=2.32"
pyjwt = ">=2.8"
psycopg2-binary = ">=2.9.9"
gunicorn = ">=22.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2"
httpx = ">=0.27"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.poetry.scripts]
# convenient launcher: `poetry run wovenlight`
wovenlight = "uvicorn:main"
# Note: we want to call uvicorn with module: `app.main:app`. Poetry scripts don't pass args here,
# so prefer the CLI shown below. Keeping script for completeness.

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
